% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/FindPlayerSeason.R
\name{FindPlayerSeason}
\alias{FindPlayerSeason}
\title{Player Season Finder}
\usage{
FindPlayerSeason(type = "advanced", year.min = "", season.start = "",
  season.end = "", c1.stat = "", c1.val = "", order.by = "")
}
\arguments{
\item{type}{currently just 'advanced'}

\item{year.min}{Earliest year to search}

\item{season.start}{Earliest year of players' careers to search}

\item{season.end}{Latest year of players' careers to search}

\item{c1.stat}{Statistic for criteria 1}

\item{c1.val}{Value for criteria 1}

\item{order.by}{Statistic to order results by}
}
\value{
data frame with player seasons that fit criteria
}
\description{
Player Season Finder
}
\examples{
FindPlayerSeason(type = 'advanced', year.min = 1980, season.start = 1, season.end = 1, c1.stat = 'mp_per_g', c1.val = 24, order.by = 'ts_pct')
}
\keyword{finder}
\keyword{player}
\keyword{season}

