% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAttendance.R
\name{GetAttendance}
\alias{GetAttendance}
\title{Attendance}
\usage{
GetAttendance(year)
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}
}
\value{
data frame with home, road, and overall attendance for each team
}
\description{
Attendance
}
\examples{
GetAttendance(2014)
}
\keyword{attendance}

