% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetGameIDs.R
\name{GetGameIDs}
\alias{GetGameIDs}
\title{Get Game IDs.}
\usage{
GetGameIDs(year, season.type = "regular")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{type}{either "regular season" or "playoffs"}
}
\value{
vector of Gamd IDs
}
\description{
Get Game IDs.
}
\examples{
GetGameIDs(2014, 'regular season')
}
\keyword{gameid}

