% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGameIDs.R
\name{GetGameIDsDay}
\alias{GetGameIDsDay}
\title{Get Game IDs for a day}
\usage{
GetGameIDsDay(date)
}
\arguments{
\item{date}{target date}
}
\value{
data.frame
}
\description{
Get Game IDs for a day
}
\examples{
GetGameIDsDay('2017-01-24')
}
\keyword{Game}

