% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGameIDs.R
\name{GetGameIDsTeam}
\alias{GetGameIDsTeam}
\title{Get Game IDs for a team}
\usage{
GetGameIDsTeam(team, year, season.type)
}
\arguments{
\item{team}{target date}

\item{year}{Season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{Either "Regular Season", "Playoffs", or "Both"}
}
\value{
data.frame
}
\description{
Get Game IDs for a team
}
\examples{
GetGameIDsTeam('Houston Rockets', 2017, 'Regular Season')
}
\keyword{Game}

