% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetGameInfo.R
\name{GetGameInfo}
\alias{GetGameInfo}
\title{Get Game Info.}
\usage{
GetGameInfo(id, src = "basketball-reference", info = c("box scores"))
}
\arguments{
\item{id}{Game ID that matches the src}

\item{src}{currently only "basketball-reference"}

\item{info}{desired information ('box scores', 'team scores', 'lineups', 'four factors')}
}
\value{
list of information requested
}
\description{
Get Game Info.
}
\examples{
GetGameInfo('201004170ATL', 'basketball-reference', c('box scores'))
}
\keyword{boxscore}

