% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHustleStats.R
\name{GetHustleStats}
\alias{GetHustleStats}
\title{Hustle Stats.}
\usage{
GetHustleStats(player.or.team = "player", year = CurrentYear(),
  season.type = "Regular Season", per.mode = "Totals", quarter = 0)
}
\arguments{
\item{player.or.team}{Either 'player' or 'team'}

\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{'Regular Season' or 'Playoffs'}

\item{per.mode}{'Per Game' or 'Totals'}

\item{quarter}{Quarter number (1, 2, 3, 4, 5 for OT1, 6 for OT2, etc.)}
}
\value{
data frame with hustle stats
}
\description{
Hustle Stats.
}
\examples{
GetHustleStats(2017)
}
\keyword{hustle}
\keyword{player}
\keyword{team}

