% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetLines.R
\name{GetLines}
\alias{GetLines}
\title{Betting lines.}
\usage{
GetLines(sport = "NBA", year, type = "both")
}
\arguments{
\item{sport}{either "NBA" or "NFL"}

\item{year}{season (e.g. 2008 for the 2007-08 season)}

\item{type}{either "regular season" or "playoffs" or "both"}
}
\value{
data frame with schedule and line for each game in that season
}
\description{
Betting lines.
}
\examples{
GetLines("NBA", 2014, "playoffs")
}
\keyword{betting}
\keyword{line,}
\keyword{odds,}
\keyword{schedule,}

