% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetLines.R, R/GetSchedule.R
\name{GetLinesRange}
\alias{GetLinesRange}
\title{Betting lines (multi-year).}
\usage{
GetLinesRange(year.start, year.end, type = "both")

GetLinesRange(year.start, year.end, type = "both")
}
\arguments{
\item{year.start}{season (e.g. 2008 for 2007-08 season)}

\item{year.end}{season (e.g. 2014 for 2013-14 season)}

\item{type}{either "regular season" or "playoffs" or "both"}

\item{sport}{either "NBA" or "NFL"}

\item{year.start}{season (e.g. 2008 for 2007-08 season)}

\item{year.end}{season (e.g. 2014 for 2013-14 season)}

\item{type}{either "regular season" or "playoffs" or "both"}
}
\value{
data frame with schedule and line for each game in all the seasons

data frame with schedule and results for each game in all the seasons
}
\description{
Betting lines (multi-year).

Schedule and results (multi-year).
}
\examples{
GetLinesRange("NBA", 2012, 2015, "playoffs")
GetLinesRange(2012, 2015, "playoffs")
}
\keyword{betting}
\keyword{line,}
\keyword{odds,}
\keyword{schedul}
\keyword{schedule,}

