% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLineupStats.R
\name{GetLineupStats}
\alias{GetLineupStats}
\title{Lineup stats}
\usage{
GetLineupStats(year, players = 5, type = "Advanced", date.from, date.to)
}
\arguments{
\item{year}{e.g. 2015 for 2014-15 season}

\item{players}{Number of players to get lineup stats on (2-5)}

\item{type}{'Advanced'}

\item{date.from}{starting date}

\item{date.to}{ending date}
}
\value{
data frame of stats
}
\description{
Lineup stats
}
\examples{
GetLineupStats(year = 2016, players = 5, type = 'Advanced')
}
\keyword{lineup}
\keyword{stats}

