% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMoments.R
\name{GetMoments}
\alias{GetMoments}
\title{Get Moments}
\usage{
GetMoments(game.id, event.id, keep.ids = FALSE)
}
\arguments{
\item{game.id}{Game ID}

\item{event.id}{Event ID from the game}
}
\value{
data frame of moments
}
\description{
Get Moments
}
\examples{
GetMoments('0021500100', '7')
}
\keyword{moment}

