% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOnOffStats.R
\name{GetOnOffStats}
\alias{GetOnOffStats}
\title{On-Off stats for each player on a team}
\usage{
GetOnOffStats(id, year, stat = "Advanced")
}
\arguments{
\item{id}{team's ID}

\item{stat}{which stat ('Advanced')}
}
\value{
data frame of stats
}
\description{
On-Off stats for each player on a team
}
\examples{
GetOnOffStats(id = '1610612745', stat = 'Advanced')
}
\keyword{onoff}
\keyword{team}

