% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPlayByPlay.R
\name{GetPlayByPlay}
\alias{GetPlayByPlay}
\title{NBA Player By Play}
\usage{
GetPlayByPlay(id, source = "NBA")
}
\arguments{
\item{id}{Game ID}

\item{source}{Currently only 'NBA'}
}
\value{
data frame of plays
}
\description{
NBA Player By Play
}
\examples{
GetPlayByPlay(id = '0041500317')
}
\keyword{by}
\keyword{play}

