% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPlayerClutchStats.R
\name{GetPlayerClutchStats}
\alias{GetPlayerClutchStats}
\title{Player clutch stats}
\usage{
GetPlayerClutchStats(year = CurrentYear(), season.type = "Regular Season",
  measure.type = "Advanced", per.mode = "Per Game", position = "")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{Either 'Regular Season' or 'Playoffs'}

\item{measure.type}{Either 'Basic', 'Advanced', or 'RPM'}

\item{per.mode}{Either 'Per Game', 'Totals', or '100 Possessions'}

\item{position}{Either 'G', 'F', or 'C'}
}
\value{
data frame with players stats
}
\description{
Player clutch stats
}
\examples{
GetPlayerClutchStats(2014)
}
\keyword{player}

