% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetPlayerDefenseStats.R
\name{GetPlayerDefenseStats}
\alias{GetPlayerDefenseStats}
\title{Defensive stats for players}
\usage{
GetPlayerDefenseStats(year, stat = "Greater Than 15Ft", per.mode = "Totals")
}
\arguments{
\item{year}{e.g. 2015 for 2014-15 season}

\item{stat}{which stat ('Greater Than 15Ft')}

\item{per.mode}{'Per Game' or 'Totals}
}
\value{
data frame of stats
}
\description{
Defensive stats for players
}
\examples{
GetPlayerDefenseStats(stat = 'Greater Than 15Ft')
}
\keyword{defense}
\keyword{player}

