% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPlayerIDs.R
\name{GetPlayerIDs}
\alias{GetPlayerIDs}
\title{Player IDs on websites}
\usage{
GetPlayerIDs(year = CurrentYear(), source = "NBA")
}
\arguments{
\item{year}{NBA season for which you want player IDs (e.g. 2008 for the 2007-08 season)}

\item{source}{Either 'NBA' or 'Basketball-Reference'}
}
\value{
data frame of names and IDs
}
\description{
Player IDs on websites
}
\examples{
GetPlayerIDs(2008)
}
\keyword{IDs}
\keyword{player}

