% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPlayerPicture.R
\name{GetPlayerPicture}
\alias{GetPlayerPicture}
\title{Get Player Picture}
\usage{
GetPlayerPicture(player, player.ids = NA)
}
\arguments{
\item{player}{Player's name or ID}

\item{player.ids}{Optional, data frame of player.ids}
}
\value{
player's picture
}
\description{
Get Player Picture
}
\examples{
GetPlayerPicture('201935', source = 'NBA')
}
\keyword{picture}
\keyword{player}

