% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPlayerStats.R
\name{GetPlayerStats}
\alias{GetPlayerStats}
\title{Player stats per season.}
\usage{
GetPlayerStats(year, type = "Basic", per.mode = "Per Game",
  source = "Basketball-Reference")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{type}{Either 'Basic', 'Advanced', or 'RPM'}

\item{per.mode}{Either 'Per Game', 'Totals', or '100 Possessions'}

\item{source}{Either 'Basketball-Reference', 'NBA', or 'ESPN'}
}
\value{
data frame with players stats
}
\description{
Player stats per season.
}
\examples{
GetPlayerStats(2014, 'Basic')
}
\keyword{player}

