% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPlayerStatsByTeam.R
\name{GetPlayerStatsByTeam}
\alias{GetPlayerStatsByTeam}
\title{Player stats by team}
\usage{
GetPlayerStatsByTeam(year = CurrentYear(), team.id,
  season.type = "Regular Season", measure.type = "Base",
  per.mode = "Per Game", source = "NBA")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{Either 'Regular Season' or 'Playoffs'}

\item{measure.type}{Either 'Base', 'Advanced', 'Misc', 'Four Factors', 'Scoring', 'Opponent', 'Usage', or 'Defense'}

\item{per.mode}{Either 'Totals', 'PerGame', 'MinutesPer', 'Per48', 'Per40', 'PerMinute', 'PerPossession', 'PerPlay', 'Per100Possessions', or 'Per100Plays'}

\item{source}{Currently only 'NBA'}
}
\value{
data frame with players stats
}
\description{
Player stats by team
}
\examples{
GetPlayerStatsByTeam(year = 2016, team.id = '1610612745')
}
\keyword{player}

