% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRoster.R
\name{GetRoster}
\alias{GetRoster}
\title{Team roster}
\usage{
GetRoster(year = CurrentYear(), team, team.ids)
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{team}{Either team name ('Golden State') or team id}

\item{team.ids}{Teams and their IDs}
}
\value{
data frame with player information
}
\description{
Team roster
}
\examples{
GetRoster(2014, 'Golden State')
}
\keyword{roster}

