% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSchedule.R
\name{GetSchedule}
\alias{GetSchedule}
\title{Schedule and Results.}
\usage{
GetSchedule(year = CurrentYear(), season.type = "Regular Season")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{'Regular Season' or 'Playoffs'}
}
\value{
data frame with game logs
}
\description{
Schedule and Results.
}
\examples{
GetSchedule(year = 2014, season.type = 'playoffs')
}
\keyword{game}
\keyword{logs}

