% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSchedule.R
\name{GetScheduleRange}
\alias{GetScheduleRange}
\title{Schedule and results (multi-year).}
\usage{
GetScheduleRange(sport = "NBA", year.start, year.end, season.type = "Both",
  info = "scores")
}
\arguments{
\item{sport}{'NBA' or 'NCAAB'}

\item{year.start}{Season (e.g. 2008 for 2007-08 season)}

\item{year.end}{Season (e.g. 2014 for 2013-14 season)}

\item{season.type}{Either 'Regular Season', 'Playoffs', or 'Both'}

\item{info}{Either 'scores' or 'advanced'}
}
\value{
data frame with schedule and results for each game in all the seasons
}
\description{
Schedule and results (multi-year).
}
\examples{
GetScheduleRange(2012, 2015, 'playoffs')
}
\keyword{schedule}

