% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetShootingStats.R
\name{GetShootingStats}
\alias{GetShootingStats}
\title{Shooting stats for players}
\usage{
GetShootingStats(player, year = CurrentYear(), stat = "Shot Type Detail",
  per.mode = "Totals", player.ids = NA)
}
\arguments{
\item{player}{player's ID}

\item{stat}{which stat ('Shot Distance (5ft)', 'Shot Distance (8ft)', 'Shot Area', 
'Assisted Shot', 'Shot Type Summary', 'Shot Type Detail', 'Assisted By')}

\item{player.ids}{Optional dataframe of player.ids}
}
\value{
data frame of stats
}
\description{
Shooting stats for players
}
\examples{
GetShootingStats(player = '201147', stat = 'Shot Type Detail')
}
\keyword{player}
\keyword{shooting}

