% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetShotDashboard.R
\name{GetShotDashboard}
\alias{GetShotDashboard}
\title{Shot Dashboard stats on players or teams}
\usage{
GetShotDashboard(stat, type = "team", id, opponent.id = "0")
}
\arguments{
\item{stat}{statistic to pull (e.g. 'Defender Distance')}

\item{type}{either 'player' or 'team'}

\item{id}{either player's ID or team's ID}
}
\value{
data frame of stats
}
\description{
Shot Dashboard stats on players or teams
}
\examples{
GetShotDashboard('Defender Distance', 'team')
}
\keyword{player}
\keyword{shooting}
\keyword{team}

