% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetShots.R
\name{GetShots}
\alias{GetShots}
\title{Get Shots}
\usage{
GetShots(player.id, year, season.type = "regular")
}
\arguments{
\item{player.id}{Player ID}

\item{year}{e.g. 2015 for 2014-15 season}

\item{season.type}{either 'regular', 'playoffs', or 'both'}
}
\value{
data frame of shots
}
\description{
Get Shots
}
\examples{
GetShots('201935', 2015, 'both')
}
\keyword{shots}

