% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStandings.R
\name{GetStandings}
\alias{GetStandings}
\title{Standings.}
\usage{
GetStandings(year)
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}
}
\value{
data frame with wins and losses for that season
}
\description{
Standings.
}
\examples{
GetStandings(2014)
}
\keyword{schedule}

