% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTeamClutchStats.R
\name{GetTeamClutchStats}
\alias{GetTeamClutchStats}
\title{Team Clutch Stats}
\usage{
GetTeamClutchStats(year = CurrentYear(), measure.type = "Basic",
  season.type = "Regular Season", per.mode = "Totals", quarter = 0)
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{measure.type}{Either 'Basic' or 'Advanced'}

\item{season.type}{'Regular Season' or 'Playoffs'}

\item{per.mode}{'Per Game' or 'Totals'}
}
\value{
data frame with team clutch stats
}
\description{
Team Clutch Stats
}
\examples{
GetTeamClutchStats(2014)
}
\keyword{team}

