% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTeamHustleStats.R
\name{GetTeamHustleStats}
\alias{GetTeamHustleStats}
\title{Team Hustle Stats.}
\usage{
GetTeamHustleStats(year = CurrentYear(), season.type = "Regular Season",
  per.mode = "Totals", quarter = 0)
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{'Regular Season' or 'Playoffs'}

\item{per.mode}{'Per Game' or 'Totals'}

\item{quarter}{Quarter number (1, 2, 3, 4, 5 for OT1, 6 for OT2, etc.)}
}
\value{
data frame with team stats
}
\description{
Team Hustle Stats.
}
\examples{
GetTeamHustleStats(2017)
}
\keyword{hustle}
\keyword{team}

