% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTeamIDs.R
\name{GetTeamIDs}
\alias{GetTeamIDs}
\title{Team IDs on websites}
\usage{
GetTeamIDs(year = CurrentYear(), source = "NBA")
}
\arguments{
\item{year}{Season for which you want team IDs (e.g. 2008 for the 2007-08 season)}

\item{source}{Either 'NBA' or 'ESPN'}
}
\value{
data frame of names and IDs
}
\description{
Team IDs on websites
}
\examples{
GetTeamIDs(2015)
}
\keyword{IDs}
\keyword{team}

