% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTeamPicture.R
\name{GetTeamPicture}
\alias{GetTeamPicture}
\title{Get Team Picture}
\usage{
GetTeamPicture(team.id)
}
\arguments{
\item{team.id}{Team ID from ESPN (e.g. 'hou')}
}
\value{
team's picture
}
\description{
Get Team Picture
}
\examples{
GetTeamPicture('hou')
}
\keyword{picture}
\keyword{team}

