% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPlayerShootingStats.R, R/GetTeamShootingStats.R
\name{GetTeamShootingStats}
\alias{GetTeamShootingStats}
\title{Player Shooting Stats.}
\usage{
GetTeamShootingStats(year = CurrentYear(), close.def.dist = "",
  season.type = "Regular Season")

GetTeamShootingStats(year = CurrentYear(), close.def.dist = "",
  season.type = "Regular Season")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{close.def.dist}{Either '', '6+ Feet - Wide Open'}

\item{season.type}{'Regular Season' or 'Playoffs'}

\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{close.def.dist}{Either '', '6+ Feet - Wide Open'}

\item{season.type}{'Regular Season' or 'Playoffs'}
}
\value{
data frame with shooting stats

data frame with wins and losses for that season
}
\description{
Player Shooting Stats.

Team Shooting Stats.
}
\examples{
GetPlayerShootingStats(2014)
GetTeamShootingStats(2014)
}
\keyword{player}
\keyword{shooting}
\keyword{team}

