% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTeamSpecificStats.R
\name{GetTeamSpecificStats}
\alias{GetTeamSpecificStats}
\title{Team specific stats in a season.}
\usage{
GetTeamSpecificStats(team, stat, year = as.numeric(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{team}{Team ID (e.g. "HOU")}

\item{stat}{Which stat do you want (e.g. "advanced gamelog", "roster")}

\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}
}
\value{
data frame with team's stats
}
\description{
Team specific stats in a season.
}
\examples{
GetTeamSpecificStats('GSW', 2015, 'advanced gamelog')
}
\keyword{team}

