% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTeamStats.R
\name{GetTeamStats}
\alias{GetTeamStats}
\title{Team Stats.}
\usage{
GetTeamStats(year = CurrentYear(), measure.type = "Base",
  season.type = "Regular Season", keep.average = FALSE, source = "NBA")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{measure.type}{Either 'Regular', 'Advanced', 'Team', 'Opponent', 'Shooting', 'Shooting Opponent', 
'Standings' for Basketball-Reference. Either 'Base', 'Advanced', 'Misc', 'Four Factors', 'Scoring',
'Opponent', 'Usage', or 'Defense' for NBA.}

\item{season.type}{'Regular Season' or 'Playoffs'}

\item{keep.average}{TRUE or FALSE, depending on whether you want to keep League Average info}

\item{source}{'Basketball-Reference' or 'NBA'}
}
\value{
data frame with wins and losses for that season
}
\description{
Team Stats.
}
\examples{
GetTeamStats(2014)
}
\keyword{team}

