% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTeamStatsByGame.R
\name{GetTeamStatsByGame}
\alias{GetTeamStatsByGame}
\title{Team stats by game.}
\usage{
GetTeamStatsByGame(year, season.type = "regular")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{Either 'regular' or 'playoffs'}
}
\value{
data frame with stats by game
}
\description{
Team stats by game.
}
\examples{
GetTeamStatsByGame(2014, 'regular')
}
\keyword{game}
\keyword{stats}
\keyword{team}

