% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetTrackingStats.R
\name{GetTrackingStats}
\alias{GetTrackingStats}
\title{NBA Player Tracking}
\usage{
GetTrackingStats(stat)
}
\arguments{
\item{stat}{statistic to pull (e.g. 'Passing', 'Possessions')}
}
\value{
data frame of stats
}
\description{
NBA Player Tracking
}
\examples{
GetTrackingStats('Postup')
}
\keyword{synergy}

