% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPlayerSeason.R
\name{FindPlayerSeason}
\alias{FindPlayerSeason}
\title{Player Season Finder}
\usage{
FindPlayerSeason(type = "Advanced", year.min = "", season.start = 1,
  season.end = -1, c1.stat = "", c1.val = "", c2.stat = "",
  c2.val = "", c3.stat = "", c3.val = "", order.by = "")
}
\arguments{
\item{type}{either 'Advanced' or 'Per Game'}

\item{year.min}{Earliest year to search}

\item{season.start}{Earliest year of players' careers to search}

\item{season.end}{Latest year of players' careers to search}

\item{c1.stat}{Statistic for criteria 1}

\item{c1.val}{Value for criteria 1}

\item{c2.stat}{Statistic for criteria 2}

\item{c2.val}{Value for criteria 2}

\item{c3.stat}{Statistic for criteria 3}

\item{c3.val}{Value for criteria 3}

\item{order.by}{Statistic to order results by}
}
\value{
data frame with player seasons that fit criteria
}
\description{
Player Season Finder
}
\examples{
FindPlayerSeason(type = 'Advanced', year.min = 1980, season.start = 1, season.end = 1, c1.stat = 'mp_per_g', c1.val = 24, order.by = 'ts_pct')
}
\keyword{finder}
\keyword{player}
\keyword{season}

