% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGameIDs.R
\name{GetGameIDs}
\alias{GetGameIDs}
\title{Get Game IDs.}
\usage{
GetGameIDs(year, source = "Basketball-Reference", season.type = "regular")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{source}{either 'Basketball-Reference' or 'NBA'}

\item{type}{either "regular season" or "playoffs"}
}
\value{
vector if from 'Basketball-Reference' or data frame if from 'NBA'
}
\description{
Get Game IDs.
}
\examples{
GetGameIDs(2014, 'regular season')
}
\keyword{gameid}

