% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGameIDs.R
\name{GetGameIDs}
\alias{GetGameIDs}
\title{Get Game IDs.}
\usage{
GetGameIDs(year = .CurrentYear(), source = "Basketball-Reference",
  season.type = "Regular Season")
}
\arguments{
\item{year}{Season (e.g. 2008 for the 2007-08 season)}

\item{source}{Either 'Basketball-Reference' or 'NBA'}

\item{type}{Either "Regular Season", "Playoffs", or "Both"}
}
\value{
vector if from 'Basketball-Reference' or data frame if from 'NBA'
}
\description{
Get Game IDs.
}
\examples{
GetGameIDs(2014, 'regular season')
}
\keyword{gameid}

