% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetGameInfo.R
\name{GetGameInfo}
\alias{GetGameInfo}
\title{Get Game Info.}
\usage{
GetGameInfo(id, source = "Basketball-Reference", info = c("box scores"))
}
\arguments{
\item{id}{Game ID that matches the source}

\item{source}{either 'Basketball-Reference' or 'NBA'}

\item{info}{desired information ('box scores', 'team scores', 'lineups', 'four factors', 'play by play')}
}
\value{
list of information requested
}
\description{
Get Game Info.
}
\examples{
GetGameInfo('201004170ATL', 'Basketball-Reference', c('box scores'))
}
\keyword{boxscore}

