% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGameInfo.R
\name{GetGameInfo}
\alias{GetGameInfo}
\title{Get Game Info.}
\usage{
GetGameInfo(id, info = c("box score"))
}
\arguments{
\item{id}{Game ID that matches the source}

\item{info}{desired information ('box scores', 'summary, 'lineups', 'play by play', 'advanced')}
}
\value{
list of information requested
}
\description{
Get Game Info.
}
\examples{
GetGameInfo('201004170ATL', c('box score'))
}
\keyword{boxscore}

