% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGameLog.R
\name{GetGameLog}
\alias{GetGameLog}
\title{Get individual game stats.}
\usage{
GetGameLog(player, year = as.numeric(format(Sys.Date(), "\%Y")),
  season.type = "Regular Season", player.ids)
}
\arguments{
\item{player}{NBA player (e.g. "Harden, James")}

\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{either 'Preseason', 'Regular Season', or 'Playoffs'}

\item{player.ids}{dataframe with player ids if scraping for multiple players}
}
\value{
data frame with player's stats
}
\description{
Get individual game stats.
}
\examples{
GetGameLog('Harden, James', 2015)
}
\keyword{gamelog}
\keyword{player}

