% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLines.R
\name{GetLines}
\alias{GetLines}
\title{Betting lines.}
\usage{
GetLines(sport = "NBA", year, type = "Both")
}
\arguments{
\item{sport}{either "NBA", "NFL", or "WNBA"}

\item{year}{season (e.g. 2008 for 2007-08 season)}

\item{type}{either "Regular Season" or "Playoffs" or "Both"}
}
\value{
data frame with schedule and line for each game in that season
}
\description{
Betting lines.
}
\examples{
GetLines("NBA", 2014, "playoffs")
}
\keyword{betting}
\keyword{line,}
\keyword{odds,}
\keyword{schedule,}

