% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLines.R
\name{GetLinesRange}
\alias{GetLinesRange}
\title{Betting lines (multi-year).}
\usage{
GetLinesRange(sport = "NBA", year.start, year.end, type = "both")
}
\arguments{
\item{sport}{either "NBA" or "NFL"}

\item{year.start}{season (e.g. 2008 for 2007-08 season)}

\item{year.end}{season (e.g. 2014 for 2013-14 season)}

\item{type}{either "regular season" or "playoffs" or "both"}
}
\value{
data frame with schedule and line for each game in all the seasons
}
\description{
Betting lines (multi-year).
}
\examples{
GetLinesRange("NBA", 2012, 2015, "playoffs")
}
\keyword{betting}
\keyword{line,}
\keyword{odds,}
\keyword{schedule,}

