% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOnOffStats.R
\name{GetOnOffStats}
\alias{GetOnOffStats}
\title{On-Off stats for each player on a team}
\usage{
GetOnOffStats(team, year = CurrentYear(), season.type = "Regular Season",
  measure.type = "Base", per.mode = "Totals", team.ids)
}
\arguments{
\item{team}{Either team name ('Golden State') or team id}

\item{year}{2015 for 2014-15 season}

\item{season.type}{Either 'Regular Season' or 'Playoffs'}

\item{measure.type}{Either 'Base', 'Advanced', 'Misc', 'Four Factors', 'Scoring', 'Opponent', 'Usage', or 'Defense'}

\item{per.mode}{Either 'Totals', 'PerGame', 'MinutesPer', 'Per48', 'Per40', 'PerMinute', 'PerPossession', 'PerPlay', 'Per100Possessions', or 'Per100Plays'}

\item{team.ids}{Teams and their IDs}
}
\value{
data frame of stats
}
\description{
On-Off stats for each player on a team
}
\examples{
GetOnOffStats(team = '1610612745', measure.type = 'Advanced')
}
\keyword{onoff}
\keyword{team}

