% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOnOffStats.R
\name{GetOnOffStats}
\alias{GetOnOffStats}
\title{On-Off stats for each player on a team}
\usage{
GetOnOffStats(id, year = .CurrentYear(), season.type = "Regular Season",
  measure.type = "Base")
}
\arguments{
\item{id}{team's ID}

\item{year}{2015 for 2014-15 season}

\item{season.type}{Either 'Regular Season' or 'Playoffs'}

\item{measure.type}{Either 'Base', 'Advanced', 'Misc', 'Four Factors', 'Scoring', 'Opponent', 'Usage', or 'Defense'}
}
\value{
data frame of stats
}
\description{
On-Off stats for each player on a team
}
\examples{
GetOnOffStats(id = '1610612745', measure.type = 'Advanced')
}
\keyword{onoff}
\keyword{team}

