% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPlayerPicture.R
\name{GetPlayerPicture}
\alias{GetPlayerPicture}
\title{Get Player Picture}
\usage{
GetPlayerPicture(player.id, source = "NBA")
}
\arguments{
\item{player.id}{Player ID}

\item{source}{currently only 'NBA'}
}
\value{
player's picture
}
\description{
Get Player Picture
}
\examples{
GetPlayerPicture('201935', source = 'NBA')
}
\keyword{picture}
\keyword{player}

