% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPlayerSpecificStats.R
\name{GetPlayerSpecificStats}
\alias{GetPlayerSpecificStats}
\title{Player specific stats in a season.}
\usage{
GetPlayerSpecificStats(player, measure.type, per.mode = "Totals",
  year = CurrentYear(), season.type = "Regular Season", player.ids = NA)
}
\arguments{
\item{player}{NBA player (e.g. "James Harden")}

\item{measure.type}{Either 'Base', 'Advanced', 'Misc', 'Scoring', or 'Usage'}

\item{per.mode}{Either 'Totals', 'PerGame'}

\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{Either 'Regular Season' or 'Playoffs'}

\item{player.ids}{Players and their IDs}
}
\value{
data frame with player's stats
}
\description{
Player specific stats in a season.
}
\examples{
GetPlayerSpecificStats('Anthony Davis', 'on-Scoring', 2015)
}
\keyword{player}

