% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPlayerSpecificStats.R
\name{GetPlayerSpecificStats}
\alias{GetPlayerSpecificStats}
\title{Player specific stats in a season.}
\usage{
GetPlayerSpecificStats(player, stat, year = as.numeric(format(Sys.Date(),
  "\%Y")), player.ids)
}
\arguments{
\item{player}{NBA player (e.g. "James Harden")}

\item{stat}{Which stat do you want (e.g. "on-off")}

\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}
}
\value{
data frame with player's stats
}
\description{
Player specific stats in a season.
}
\examples{
GetPlayerSpecificStats('Anthony Davis', 'on-off', 2015)
}
\keyword{player}

