% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetPlayerSpecificStats.R
\name{GetPlayerSpecificStats}
\alias{GetPlayerSpecificStats}
\title{Player's stat in a season.}
\usage{
GetPlayerSpecificStats(player, year, stat)
}
\arguments{
\item{player}{NBA player (e.g. "James Harden")}

\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{stat}{Which stat do you want (e.g. "on-off")}
}
\value{
data frame with player's stats
}
\description{
Player's stat in a season.
}
\examples{
GetPlayerSpecificStats('Anthony Davis', 2015, 'on-off)
}
\keyword{player}

