% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPlayerStats.R
\name{GetPlayerStats}
\alias{GetPlayerStats}
\title{Player stats per season.}
\usage{
GetPlayerStats(year = .CurrentYear(), season.type = "Regular Season",
  measure.type = "Basic", per.mode = "Per Game", position = "",
  source = "Basketball-Reference")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{Either 'Regular Season' or 'Playoffs'}

\item{measure.type}{Either 'Basic', 'Advanced', or 'RPM'}

\item{per.mode}{Either 'Per Game', 'Totals', or '100 Possessions'}

\item{position}{Either 'Guard', 'Forward', or 'Center'}

\item{source}{Either 'Basketball-Reference', 'NBA', or 'ESPN'}
}
\value{
data frame with players stats
}
\description{
Player stats per season.
}
\examples{
GetPlayerStats(2014, measure.type = 'Basic')
}
\keyword{player}

