% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetPlayerStats.R
\name{GetPlayerStats}
\alias{GetPlayerStats}
\title{Player stats per season.}
\usage{
GetPlayerStats(year, type = "totals")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{type}{Either 'per game', 'totals', or 'advanced'}
}
\value{
data frame with players stats
}
\description{
Player stats per season.
}
\examples{
GetPlayerStats(2014, 'totals')
}
\keyword{player}

