% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetSchedule.R
\name{GetSchedule}
\alias{GetSchedule}
\title{Schedule and Results.}
\usage{
GetSchedule(year, season.type = "regular")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{either 'regular', 'playoffs', or 'both'}
}
\value{
data frame with schedule and results for each game in that season
}
\description{
Schedule and Results.
}
\examples{
GetSchedule(2014, 'playoffs')
}
\keyword{schedule}

