% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSchedule.R
\name{GetSchedule}
\alias{GetSchedule}
\title{Schedule and Results.}
\usage{
GetSchedule(sport = "nba", year, season.type = "regular", info = "scores")
}
\arguments{
\item{sport}{'NBA' or 'NCAAB'}

\item{year}{season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{either 'regular', 'playoffs', or 'both'}

\item{info}{either 'scores' or 'advanced'}
}
\value{
data frame with schedule and results for each game in that season
}
\description{
Schedule and Results.
}
\examples{
GetSchedule(2014, 'playoffs')
}
\keyword{schedule}

