% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSchedule.R
\name{GetSchedule}
\alias{GetSchedule}
\title{Schedule and Results.}
\usage{
GetSchedule(sport = "NBA", year = .CurrentYear(),
  season.type = "Regular Season", info = "scores")
}
\arguments{
\item{sport}{'NBA' or 'NCAAB'}

\item{year}{Season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{Either 'Regular Season', 'Playoffs', or 'Both'}

\item{info}{Either 'scores' or 'advanced'}
}
\value{
data frame with schedule and results for each game in that season
}
\description{
Schedule and Results.
}
\examples{
GetSchedule(year = 2014, season.type = 'playoffs')
}
\keyword{schedule}

