% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetSchedule.R
\name{GetScheduleRange}
\alias{GetScheduleRange}
\title{Schedule and results (multi-year).}
\usage{
GetScheduleRange(sport = "NBA", year.start, year.end, season.type = "both",
  info = "scores")
}
\arguments{
\item{sport}{'NBA' or 'NCAAB'}

\item{year.start}{season (e.g. 2008 for 2007-08 season)}

\item{year.end}{season (e.g. 2014 for 2013-14 season)}

\item{season.type}{either 'regular' or 'playoffs' or 'both'}

\item{info}{either 'scores' or 'advanced'}
}
\value{
data frame with schedule and results for each game in all the seasons
}
\description{
Schedule and results (multi-year).
}
\examples{
GetScheduleRange(2012, 2015, 'playoffs')
}
\keyword{schedule}

