% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetShootingStats.R
\name{GetShootingStats}
\alias{GetShootingStats}
\title{Shooting stats for players}
\usage{
GetShootingStats(id, stat = "Shot Type Detail")
}
\arguments{
\item{id}{player's ID}

\item{stat}{which stat ('Shot Type Detail')}
}
\value{
data frame of stats
}
\description{
Shooting stats for players
}
\examples{
GetShootingStats(id = '201147', stat = 'Shot Type Detail')
}
\keyword{player}
\keyword{shooting}

