% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetSynergyStats.R
\name{GetSynergyStats}
\alias{GetSynergyStats}
\title{Synergy stats on players or teams}
\usage{
GetSynergyStats(stat, side = "offensive", type = "player")
}
\arguments{
\item{stat}{statistic to pull (e.g. 'Postup', 'Isolation', 'PRRollMan')}

\item{side}{either 'offensive' or 'defensive'}

\item{type}{either 'player' or 'team'}
}
\value{
data frame of stats
}
\description{
Synergy stats on players or teams
}
\examples{
GetSynergyStats('Postup')
}
\keyword{synergy}

