% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetTeamIDs.R
\name{GetTeamIDs}
\alias{GetTeamIDs}
\title{Team IDs on websites}
\usage{
GetTeamIDs(year = 2016, source = "NBA")
}
\arguments{
\item{year}{NBA season for which you want team IDs (e.g. 2008 for the 2007-08 season)}

\item{source}{website that is being used ('NBA', 'ESPN', 'Basketball-Reference')}
}
\value{
data frame of names and IDs
}
\description{
Team IDs on websites
}
\examples{
GetTeamIDs(2015)
}
\keyword{IDs}
\keyword{team}

