% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTeamPicture.R
\name{GetTeamPicture}
\alias{GetTeamPicture}
\title{Get Team Picture}
\usage{
GetTeamPicture(team.id, source = "ESPN")
}
\arguments{
\item{team.id}{Team ID}

\item{source}{currently only 'ESPN'}
}
\value{
team's picture
}
\description{
Get Team Picture
}
\examples{
GetTeamPicture('hou', source = 'ESPN')
}
\keyword{picture}
\keyword{team}

