% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTeamShootingStats.R
\name{GetTeamShootingStats}
\alias{GetTeamShootingStats}
\title{Team Shooting Stats.}
\usage{
GetTeamShootingStats(year = .CurrentYear(), close.def.dist = "",
  season.type = "Regular Season")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{close.def.dist}{Either '', '6+ Feet - Wide Open'}

\item{season.type}{'Regular Season' or 'Playoffs'}
}
\value{
data frame with wins and losses for that season
}
\description{
Team Shooting Stats.
}
\examples{
GetTeamShootingStats(2014)
}
\keyword{shooting}
\keyword{team}

