% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GetTeamStats.R
\name{GetTeamStats}
\alias{GetTeamStats}
\title{Team Stats.}
\usage{
GetTeamStats(year, stat.type = "regular", season.type = "regular")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{season.type}{'regular' or 'playoffs'}
}
\value{
data frame with wins and losses for that season
}
\description{
Team Stats.
}
\examples{
GetTeamStats(2014)
}
\keyword{team}

