% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTeamStats.R
\name{GetTeamStats}
\alias{GetTeamStats}
\title{Team Stats.}
\usage{
GetTeamStats(year = CurrentYear(), measure.type = "Base",
  season.type = "Regular Season", per.mode = "Totals", quarter = 0)
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{measure.type}{Either 'Base', 'Advanced', 'Misc', 'Four Factors', 'Scoring',
'Opponent', 'Usage', or 'Defense'}

\item{season.type}{'Regular Season' or 'Playoffs'}

\item{per.mode}{'Per Game' or 'Totals'}

\item{quarter}{Quarter number (1, 2, 3, 4, 5 for OT1, 6 for OT2, etc.)}
}
\value{
data frame with team stats
}
\description{
Team Stats.
}
\examples{
GetTeamStats(2014)
}
\keyword{team}

