% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTeamStats.R
\name{GetTeamStats}
\alias{GetTeamStats}
\title{Team Stats.}
\usage{
GetTeamStats(year = .CurrentYear(), stat.type = "Regular",
  season.type = "Regular Season", keep.average = FALSE,
  source = "Basketball-Reference")
}
\arguments{
\item{year}{NBA season (e.g. 2008 for the 2007-08 season)}

\item{stat.type}{Either 'Regular', 'Advanced', 'Team', 'Opponent', 'Shooting', 'Shooting Opponent', 'Standings'}

\item{season.type}{'Regular Season' or 'Playoffs'}

\item{keep.average}{TRUE or FALSE, depending on whether you want to keep League Average info}

\item{source}{'Basketball-Reference' or 'NBA'}
}
\value{
data frame with wins and losses for that season
}
\description{
Team Stats.
}
\examples{
GetTeamStats(2014)
}
\keyword{team}

