% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTrackingStats.R
\name{GetTrackingStats}
\alias{GetTrackingStats}
\title{NBA Player Tracking}
\usage{
GetTrackingStats(year = .CurrentYear(), season.type = "Regular Season",
  per.mode = "Per Game", stat, player.or.team = "Player", position = "")
}
\arguments{
\item{year}{2015 for 2014-15 season}

\item{season.type}{Either 'Regular Season' or 'Playoffs'}

\item{per.mode}{Either 'Per Game', 'Totals', or '100 Possessions'}

\item{stat}{Statistic to pull (e.g. 'Passing', 'Possessions')}

\item{player.or.team}{Either 'Player' or 'Team'}

\item{position}{Either 'G', 'F', or 'C'}
}
\value{
data frame of stats
}
\description{
NBA Player Tracking
}
\examples{
GetTrackingStats(stat = 'Postup')
}
\keyword{synergy}

