% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTrackingStats.R
\name{GetTrackingStats}
\alias{GetTrackingStats}
\title{NBA Player Tracking}
\usage{
GetTrackingStats(year = .CurrentYear(), season.type = "Regular Season",
  per.mode = "Per Game", measure.type, player.or.team = "Player",
  position = "")
}
\arguments{
\item{year}{2015 for 2014-15 season}

\item{season.type}{Either 'Regular Season' or 'Playoffs'}

\item{per.mode}{Either 'Per Game' or 'Totals'}

\item{measure.type}{Either 'SpeedDistance', 'Rebounding', 'Possessions', 'CatchShoot', 'PullUpShot', 
'Defense', 'Drives', 'Passing', 'ElbowTouch', 'PostTouch', 'PaintTouch', or 'Efficiency'}

\item{player.or.team}{Either 'Player' or 'Team'}

\item{position}{Either 'G', 'F', 'C', 'G-F', 'F-G', 'F-C', or 'C-F'}
}
\value{
data frame of stats
}
\description{
NBA Player Tracking
}
\examples{
GetTrackingStats(stat = 'Postup')
}
\keyword{synergy}

